Revert "Merge branch 'main' into zky"
This reverts commit a9c97f314799f7290a04791a4f823663996c20fb, reversing
changes made to 9ade3457a09424711668d2299d336d45612b263e.

+++ src/main/java/project/model/Comment.java
@@ -16,7 +16,7 @@ public class Comment {
     //评论内容
     private String content;
     //评分
-    private double score;
+    private int score;
     //评论时间
     private Date create_time;
 
@@ -52,7 +52,7 @@ public class Comment {
         this.content = content;
     }
 
-    public double getScore() {
+    public int getScore() {
         return score;
     }
